pyproject.toml
[tool.poetry]
name = "rmap"
version = "0.2.4"
description = "Automated enumeration for red teamers"
authors = ["syspuke <syspuke@pm.me>"]
license = "gpl-3.0"

[tool.poetry.dependencies]
python = "^3.10"
pexpect = "^4.8.0"
colorama = "^0.4.6"
python-libnmap = "^0.7.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
